requirements.txt
numpy
matplotlib
pymysql
pandas
fpdf2
requests
datetime
sqlalchemy
plotly
psycopg2-binary